Merge from leeca03 rev 42.
Step 2 of providing RelationCount node selectors.

This cleans up the way selector types are handled in the SelectionEditorTool.
Each type of node selector is now expected to be a standard NodeSelectorPart,
with it's own UI component and a generator/provider for a node selector
instance.

Additionally, the handling of selection composition is cleaned up.  The
different composition modes are now defined by an enum type.  Future
improvements might include a ToolTip text and a clearer name (e.g.
SelectionCompositionMode).  The new default is selected + new nodes, which is
probably less surprising that the previous Extend Graph choice.

At some future time, PathMatcher should be renamed as NodeSelector.