Overhaul Java dependency analysis, adding support for packages.  This extends the DepanFileSystem TreeLoader for directories, and a revised JarFileLister for jars source.  Additionally, these analyzers have been moved around in the tree structure.
The Java relations have been changed to support the new package structure, with
packages contained in directories and classes contained in files in addition to
their containment within a package hierarhcy.

A new test directory for some of the DepanBytecode features is also included.
This structure lead to a renaming of the project for DepanCore tests.  That
project is now named TestCore.

In this rewrite, support for progress monitors and some forms of file and
package filtering has been lost.  These features should get restored in the
near future.