Initial changes to clean up DepAn persistence.  This move the serialization code for GraphModels up into the DepanApp layer.  As we move forward, the DepanCore layer will only know about manipulating dependency graphs, but not where they come from.
Serialization of GraphModels is now handles by a GraphModelConverter.  This
allows GraphModels to be embedded in other kinds of serializable objects.  This
will be a bonus as we move to richer user interfaces that know somewhat about
the kinds of dependencies graphs that are being manipulated (e.g. do they
mainly contain Java relations).