Convert DepAn to use a robust scheme for persistence of graph views.  The older ".dpanv" format, which used ZIP-based directory encoding of persisted properties is no longer supported (and is unreadable).  The new "DepAn View Information" format (".dvi") is based on XStream serialization and is written as text files.
The basis of this format is that all rendering choices are stored in a
ViewPreferences instance which can be serialized straightforwardly.  The
details about the dependency graph and the viewed nodes are stored separately.
A ViewDocument contains these three components (graph, nodes, prefs), and is
serialized using a few custom XStream Converters.