Use LayoutPickerControl in GraphEditor.
Eliminate LayoutSelector infavor of LayoutPickerControl.  The new class takes
advantages of the SWT architecture; the old one was a hack.

Add a LabeledControl composite SWT object for handling the common case of a
text label before a control.  The control itself still needs to be added, and
have its layout data set correctly.  Static factory methods in this class
should handle the normal cases.