Change the way post-analysis dependency-graph are written.  The original .dpang format is being obsoleted, with a new .dgi (DepAg Graph Info) document type taking over.  This is roughly equivalent to the previous change for dpanv files.
The DGI format provides additional information about the graph as a seperate section of the document.  Currently, the expected analysis plugins for the document are added.  The UI can take advantage of this information to present the user with a more rationale choice of options.

Additionally, cleaned up a few compilation warnings.  Mostly be adding the appropriate @SuppressWarning annotation.