Make all the relationship set selectors use the only the relationship sets that are provided by the associated plugins for each graph.  A major part of this involve configuring the RelSetPickers only when a graph is connected to them.  Setting them up once at the beginning with the full list doesn't do the right thing.
Major cleanup of the RelationshipSetPickerControl to more effectively use the Eclipse model for Viewers.  In particular, the ContentProvider should not be inserting values directly into the underlying Control.  As part of this, several of the helper methods have been moved to a new relsets package under utils.

Some "consistency" cleanup, with most RelationshipSetPickerControl instances now named relSetPicker.