Use latest JOGL 2.1.2 from JOGAMP.
Although JOGL 2.1.2 claims that the native code can be loaded from Jar files, that doesn't seem to
work in the DepAn OSGi model.  Debug launches still require setting the java.library.path variable to the location of the DLL files.

Most of the changes are renaming the JOGL context from GL to GL2.  In the future, maybe use advance OGL profiles.  Windows 7 64-bit seems to support GL4bc.

Initialize GLScene separately from construction.  This allows some resources (Textures) to be created with an active GL context in the thread which is a new requirement.

Also, use the ByteBuffer methods directly.  The BufferUtil class is no longer available.